+++ packages/pluggableWidgets/rating-native/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+- The widget has been refactored and now includes a version of react-native-vector-icons that matches Native Template. 
+
 ## [2.1.0] - 2022-1-24
 
 ### Added